infrastructure/cdk/requirements.txt
aws-cdk-lib>=2.150.0
constructs>=10.0.0
cdk-fck-nat>=1.6.0
